explicitly set multiprocess start method to fork for cross-OS consistency
The method for launching a process can be "spawn", "fork", and "forkserver".
The default on Unix is fork, and the resulting process inherits all resources
from the parent process. Conversely, the default on Mac OS X/Windows is spawn,
which results in a minimal number of resources inherited by the child process.

https://docs.python.org/3/library/multiprocessing.html#contexts-and-start-methods